#13 refactor: 뒤로가기 이벤트 처리 커스텀 훅으로 분리
